shared cpmponent Toolbar
