+++ src/pages/global-authorization/index.js
@@ -124,7 +124,7 @@ const GlobalAuthorization = () => {
           labels={labels}
           inputSearch={true}
         >
-          <Grid item sx={{ width: '350px', pt: '7px !important' }}>
+          <Grid item sx={{ width: '350px' }}>
             <ResourceComboBox
               datasetId={DataSets.MODULE}
               name='moduleId'
@@ -138,8 +138,23 @@ const GlobalAuthorization = () => {
               }}
             />
           </Grid>
-          <Grid item sx={{ pt: '7px !important' }}>
-            <Button variant='contained' onClick={() => openApplyModuleLevel()} disabled={!filters.moduleId}>
+          <Grid item>
+            <Button
+              variant='contained'
+              sx={{
+                ml: 2,
+                backgroundColor: '#231F20',
+                '&:hover': {
+                  backgroundColor: '#231F20',
+                  opacity: 0.8
+                },
+                width: 'auto',
+                height: '35px',
+                objectFit: 'contain'
+              }}
+              onClick={() => openApplyModuleLevel()}
+              disabled={!filters.moduleId}
+            >
               <Icon icon='mdi:arrow-expand-right' fontSize={20} />
             </Button>
           </Grid>
